move server to its own class
