+++ API/app.ts
@@ -4,6 +4,7 @@ import {inject} from "inversify";
 import {iocTypes} from "./ioc-types";
 import {AppConfig} from "./config/app-config";
 import {DbConnector} from "./db-connector";
+import {Server} from "./server";
 
 /**
  * The Application
@@ -11,31 +12,23 @@ import {DbConnector} from "./db-connector";
  * Initializes the server, sets the routes, makes sure everything is connected properly
  */
 export class App {
-    server;
-
-    constructor(@inject(iocTypes.AppConfig) private appConfig: AppConfig,
-                @inject(iocTypes.DbConnector) private dbConnector: DbConnector) {
-        this.server = new hapi.Server();
-
+    constructor(@inject(iocTypes.DbConnector) private dbConnector: DbConnector,
+                @inject(iocTypes.Server) private  server: Server) {
     }
 
-    public async initializeServer() {
+    public async initializeApp() {
+        console.info(`Environment: ${process.env.NODE_ENV}`);
+
+
         // connect to the db
         await this.dbConnector.connection.authenticate();
+        console.info(`Connected to database`);
 
-        // create server
-        this.setServerConnection();
-        this.server.start();
 
-        // Log success
-        console.info(`Started server at ${JSON.stringify(this.server.info.uri)}`);
-        console.info(`Environment: ${process.env.NODE_ENV}`);
+        // init server
+        await this.server.initializeServer();
+        const serverUri = JSON.stringify(this.server.server.info.uri);
+        console.info(`Started server at ${serverUri}`);
     }
 
-    private setServerConnection() {
-        this.server.connection({
-            host: this.appConfig.server.host,
-            port: this.appConfig.server.portNumber
-        });
-    }
 }
