+++ API/main.ts
@@ -4,19 +4,19 @@ import "reflect-metadata";
 import {initializeBindings, iocContainer} from "./config/inversify.config";
 import {App} from "./app";
 import {iocTypes} from "./ioc-types";
-import {AppConfig} from "./config/app-config";
 import {DbConnector} from "./db-connector";
+import {Server} from "./server";
 
 // initialize the ioc bindings so we can inject properly
 initializeBindings();
 
 // create instance of app and inject dependencies
-const appConfig = iocContainer.get<AppConfig>(iocTypes.AppConfig);
 const dbConnector = iocContainer.get<DbConnector>(iocTypes.DbConnector);
-const app = new App(appConfig, dbConnector);
+const server = iocContainer.get<Server>(iocTypes.Server);
+const app = new App(dbConnector, server);
 
 // initialize app
-app.initializeServer()
+app.initializeApp()
     .catch((err) => {
     // if there is an uncaught error log it and exit the application
         // NOTE: exiting an application is really bad but this isn't gonna happen unless something really really catastrophic happens, at which point out best bet is to kill this node process and alert the process manager running it that we did. The manager will the take care of running a new process so we have no down time
